Increase precision in Monte Carlo simulator

+++ include/simulator/MonteCarloSimulator.hpp
@@ -34,14 +34,14 @@ class MonteCarloSimulator {
         const std::string toEventColumn,
         const int binSize
     );
-    std::map<std::pair<float, float>, float> createHistogram(const std::vector<float>& data, int numBins);
-    float generateRandomWaitTimeFromHistogram(const std::map<std::pair<float, float>, float>& histogram);
+    std::map<std::pair<float, float>, double> createHistogram(const std::vector<float>& data, int numBins);
+    float generateRandomWaitTimeFromHistogram(const std::map<std::pair<float, float>, double>& histogram);
 
  public:
-    std::vector<float> hourlyIncidentProbabilityDistribution;
-    std::vector<std::vector<float>> minuteIncidentProbabilityDistribution;
-    std::map<std::pair<std::string, std::string>, std::map<std::string, std::map<std::pair<float, float>, float>>> waitTimesHistograms;
-    std::vector<std::vector<float>> triageProbabilityDistribution;
+    std::vector<double> hourlyIncidentProbabilityDistribution;
+    std::vector<std::vector<double>> minuteIncidentProbabilityDistribution;
+    std::map<std::pair<std::string, std::string>, std::map<std::string, std::map<std::pair<float, float>, double>>> waitTimesHistograms;
+    std::vector<std::vector<double>> triageProbabilityDistribution;
     std::map<int, int64_t> indexToGridIdMapping;
     std::map<int64_t, int> gridIdToIndexMapping;
     std::vector<std::vector<std::vector<double>>> locationProbabilityDistribution;
